preliminary modular support for container providers and kubernetes implementation
